+++ src/library/Identicon/index.tsx
@@ -2,34 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { useCallback } from 'react';
-import styled from 'styled-components';
-import type { Circle } from './circles';
 import { getCircleXY, OUTER_CIRCLE, renderCircle, Z } from './circles';
 import { getColors } from './colors';
-
-export interface IdenticonProps {
-  size: number;
-  disableCursorCopy?: boolean;
-  disableClipboardCopy?: boolean;
-  colors?: string[];
-  value: string;
-}
-
-const Wrapper = styled.div<{
-  $disableCursorCopy?: boolean;
-  $disableClipboardCopy?: boolean;
-}>`
-  svg > circle:first-child {
-    fill: var(--border-primary-color);
-  }
-
-  .cursor-copy {
-    cursor: ${(props) =>
-      !(props.$disableCursorCopy || props.$disableClipboardCopy)
-        ? 'copy'
-        : 'default'};
-  }
-`;
+import type { Circle, IdenticonProps } from './types';
+import { Wrapper } from './Wrapper';
 
 export const Identicon = ({
   size,
@@ -59,11 +35,7 @@ export const Identicon = ({
       $disableCursorCopy={disableCursorCopy}
       $disableClipboardCopy={disableClipboardCopy}
     >
-      <div
-        aria-hidden="true"
-        onClick={() => copyToClipboard}
-        className="cursor-copy"
-      >
+      <div aria-hidden="true" onClick={() => copyToClipboard} className="copy">
         <svg
           height={size}
           id={value}
